+++ jest.setup.js
@@ -1,0 +1,4 @@
+process.on("unhandledRejection", (reason, promise) => {
+  console.log(promise, reason, reason.stack);
+  process.exit(1);
+});