+++ src/polyfill.test.js
@@ -1,6 +1,7 @@
 import { useBrowserPolyfill } from "./polyfill";
 
-afterEach(() => {
+beforeEach(() => {
+  document.body.innerHTML = "";
   delete window.appHistory;
 });
 
@@ -34,7 +35,7 @@ describe("useBrowserPolyfill", () => {
       done();
     });
 
-    document.body.innerHTML = `<div><a href="/page"><span>Page<span></a></div>`;
+    document.body.innerHTML = `<div><a href="/page"><span>Page</span></a></div>`;
 
     document.querySelector("span").click();
   });
@@ -56,14 +57,41 @@ describe("useBrowserPolyfill", () => {
 
   it("should prevent default on anchor/area clicks so that navigation doesn't happen", async (done) => {
     useBrowserPolyfill({ configurable: true });
-    window.addEventListener("click", (evt) => {
-      expect(evt.defaultPrevented).toBe(true);
-      done();
-    });
+    window.addEventListener(
+      "click",
+      (evt) => {
+        expect(evt.defaultPrevented).toBe(true);
+        done();
+      },
+      { once: true }
+    );
 
     document.body.innerHTML = `<div><a href="/page">Page</a></div>`;
 
     document.querySelector("a").click();
+  });
+
+  it("should abort a previous anchor click if the promise isn't complete yet", async () => {
+    useBrowserPolyfill({ configurable: true });
+
+    let firstRespondWith;
+    window.appHistory.addEventListener("navigate", (evt) => {
+      if (evt.destination.url === "/page1") {
+        firstRespondWith = new Promise((resolve) => setTimeout(resolve, 10));
+        evt.respondWith(firstRespondWith);
+      }
+    });
+
+    document.body.innerHTML = `<div><a href="/page1">Page1</a><a href="/page2">Page2</a></div>`;
+
+    [...document.querySelectorAll("a")].forEach((ele) => ele.click());
+
+    await firstRespondWith;
+
+    expect(window.appHistory.entries.length).toBe(3);
+    expect(window.appHistory.current.url).toBe("http://localhost/page2");
+    expect(window.appHistory.entries[1].url).toBe("http://localhost/page1");
+    expect(window.appHistory.entries[1].finished).toBe(false);
   });
 
   it("should not error out if no param passed", () => {